refactor: update monitor dashboard visuals with new background images and enhance loading indicators
