+++ frontend/pages/monitor-dashboard/components/MonitorLeftColumn.jsx
@@ -77,21 +77,23 @@ export default function MonitorLeftColumn({ dailyTokens, instanceList, loading, 
   return (
     <div className="flex flex-col gap-3 w-full lg:w-1/4 h-[calc(100%+2.5rem)] lg:-mt-10">
       <MonitorPanel title="每日 Token 消耗" className="shrink-0">
-        {loading ? (
-          <div className="h-full min-h-[180px] flex items-center justify-center px-4">
-            <div className="text-center">
-              <div className="mx-auto mb-2 flex h-8 w-8 items-center justify-center">
-                <svg className="h-6 w-6 animate-spin text-[#5ba6d6]" fill="none" viewBox="0 0 24 24">
-                  <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-                  <path className="opacity-90" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z" />
-                </svg>
+        <div className="h-[100px] w-full min-h-0 shrink-0 overflow-hidden">
+          {loading ? (
+            <div className="flex h-full w-full items-center justify-center px-4">
+              <div className="text-center">
+                <div className="mx-auto mb-2 flex h-8 w-8 items-center justify-center">
+                  <svg className="h-6 w-6 animate-spin text-[#5ba6d6]" fill="none" viewBox="0 0 24 24">
+                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
+                    <path className="opacity-90" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z" />
+                  </svg>
+                </div>
+                <div className="text-sm text-[#8fb1c6]">加载中...</div>
               </div>
-              <div className="text-sm text-[#8fb1c6]">加载中...</div>
             </div>
-          </div>
-        ) : (
-          <ReactECharts option={getDailyTokenOption(dailyTokens)} style={{ height: "100%", width: "100%" }} />
-        )}
+          ) : (
+            <ReactECharts option={getDailyTokenOption(dailyTokens)} style={{ height: "100%", width: "100%" }} />
+          )}
+        </div>
       </MonitorPanel>
 
       <MonitorPanel
