+++ frontend/pages/monitor-dashboard/index.jsx
@@ -6,8 +6,8 @@ import { useMonitorFullscreen } from "./hooks/useMonitorFullscreen.js";
 import { useMonitorDashboard } from "./hooks/useMonitorDashboard.js";
 import { useMonitorSession } from "./hooks/useMonitorSession.js";
 import bgImage from "./images/opsRobotbg3.png";
-import topBg from "./images/centertopbg2.png";
-import topLineBg from "./images/toplinebg.png";
+import topBg from "./images/centertopbg3.png";
+import topLineBg from "./images/toplinebg1.png";
 
 export default function MonitorDashboard() {
   const { containerRef, isFullscreen, toggleFullscreen } = useMonitorFullscreen();
@@ -27,20 +27,20 @@ export default function MonitorDashboard() {
   return (
     <div
       ref={containerRef}
-      className={`w-full flex-1 text-white p-2 pt-0 font-sans selection:bg-[#00f0ff]/30 flex flex-col gap-0 relative overflow-y-auto lg:overflow-hidden ${
+      className={`w-full bg-[#010611] flex-1 text-white p-2 pt-0 font-sans selection:bg-[#00f0ff]/30 flex flex-col gap-0 relative overflow-y-auto lg:overflow-hidden ${
         isFullscreen ? "h-screen" : "min-h-[750px]"
       }`}
-      style={{
-        backgroundImage: `url(${bgImage})`,
-        backgroundSize: "cover",
-        backgroundPosition: "center",
-        backgroundRepeat: "no-repeat",
-        backgroundColor: "#010611"
-      }}
+      // style={{
+      //   backgroundImage: `url(${bgImage})`,
+      //   backgroundSize: "cover",
+      //   backgroundPosition: "center",
+      //   backgroundRepeat: "no-repeat",
+      //   backgroundColor: "#010611"
+      // }}
     >
       {/* 顶部贯穿背景线 */}
       <div 
-        className="absolute top-0 left-0 w-full h-[15px] z-9 pointer-events-none"
+        className="absolute top-0 left-0 w-full h-[24px] z-9 pointer-events-none"
         style={{
           backgroundImage: `url(${topLineBg})`,
           backgroundRepeat: "repeat-x",
